requirements.txt
apache-airflow-providers-google
requests
beautifulsoup4
pandas